+++ src/CleanRegex/Replace/Details/Group/ReplaceGroup.php
@@ -1,7 +1,9 @@
 <?php
 namespace TRegx\CleanRegex\Replace\Details\Group;
 
+use Throwable;
 use TRegx\CleanRegex\Match\Details\Group\Group;
+use TRegx\CleanRegex\Match\Optional;
 
 interface ReplaceGroup extends Group
 {
@@ -10,4 +12,24 @@ interface ReplaceGroup extends Group
     public function modifiedOffset(): int;
 
     public function byteModifiedOffset(): int;
+
+    /**
+     * @deprecated
+     */
+    public function orThrow(Throwable $throwable = null);
+
+    /**
+     * @deprecated
+     */
+    public function orReturn($substitute);
+
+    /**
+     * @deprecated
+     */
+    public function orElse(callable $substituteProducer);
+
+    /**
+     * @deprecated
+     */
+    public function map(callable $mapper): Optional;
 }